[Fix] AppData 및 DTO 내 상수 타입 변경 (#95)

+++ Going-iOS/Network/DTO/Trip/Response/CodeResponseDTO.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct CodeResponseDTO : Response, DTO {
-    let tripId: Int64
+    let tripId: Int
     let title, startDate, endDate: String
     let day: Int
 }
